Use DateTime instead of DateTimeOffset
According to the documentation, DateTimeOffset is mapped to "time with timezone" rather than a "timestamp with timezone". We expected it to be mapped to a timestamp, and that what has happened, but only because we used the CURRENT_TIMESTAMP default. To avoid any confusion we decided to migrate all DateTimeOffset fields to DateTime and always use UTC timestamps.

See https://www.npgsql.org/doc/types/datetime.html#net-types-and-postgresql-types