Add a db-dump target to the Makefile
While running the application locally with docker, I managed to bring
everything down and reset the database. It would be useful to have a
dump of the initial data for this situation, so going through user
registration etc. isn't necessary.

This addes a `db-dump` target to the Makefile so you can run something
like `make db-dump > db.sql` to get a dump of the database.